# Optimize Enumerable.slice/1 for lists and maps