@josevalim 2020-11-14
Hi @eksperimental, sorry but I don't see how these changes are improving the code. My theory is that this is actually happening because EnumPatch is bypassing the whole protocol. A more appropriate comparison would be to have EnumPatch call a EnumerablePatch protocol. Is that what you are doing? Can you please share the benchmark code?