@josevalim 2020-11-16
I see! Please update the PR with the Enumerable.List only clauses. Thank you!