@Strech 2020-11-16
@eksperimental Thanks for picking it up for `Enum`.