@eksperimental 2020-11-17
> I see! Please update the PR with the Enumerable.List only clauses. Thank you!

done